mainは完成したbufferを描画することだけに関心を持つ
